refactor : component route 변경
